Move read_formatted_frame() and write_formatted_frame()
- move from swn.modflow to swn.file
