+++ docs/source/file.rst
@@ -9,6 +9,7 @@ Methods
 .. autosummary::
    :toctree: ref/
 
-   topnet2ts
-   gdf_to_shapefile
-
+    topnet2ts
+    gdf_to_shapefile
+    read_formatted_frame
+    write_formatted_frame
