+++ swn/modflow/__init__.py
@@ -1,16 +1,10 @@
 """MODFLOW module for a surface water network."""
 from ._misc import geotransform_from_flopy
-from ._swnmf6 import (
-    SwnMf6,
-    read_formatted_frame,
-    write_formatted_frame,
-)
+from ._swnmf6 import SwnMf6
 from ._swnmodflow import SwnModflow
 
 __all__ = [
     "SwnModflow",
     "SwnMf6",
     "geotransform_from_flopy",
-    "read_formatted_frame",
-    "write_formatted_frame",
 ]
